refactor: Fixes to unit tests
