+++ src/components/ColorRampDropdown.tsx
@@ -3,7 +3,15 @@ import { Button, Tooltip } from "antd";
 import React, { ReactElement, useContext, useEffect, useMemo, useState } from "react";
 import { Color } from "three";
 
-import { CategoricalPaletteMap, ColorRamp, ColorRampData, ColorRampMap, ColorRampType } from "../colorizer";
+import {
+  CategoricalPaletteMap,
+  ColorRamp,
+  ColorRampData,
+  ColorRampMap,
+  ColorRampType,
+  DEFAULT_CATEGORICAL_PALETTES,
+  DEFAULT_COLOR_RAMPS,
+} from "../colorizer";
 import { AppThemeContext } from "./AppStyle";
 import styles from "./ColorRampDropdown.module.css";
 import IconButton from "./IconButton";
@@ -11,10 +19,10 @@ import IconButton from "./IconButton";
 type ColorRampSelectorProps = {
   selectedRamp: string;
   onChangeRamp: (colorRampKey: string, reversed: boolean) => void;
-  colorRamps: ColorRampMap;
+  colorRamps?: ColorRampMap;
 
   useCategoricalPalettes?: boolean;
-  categoricalPalettes: CategoricalPaletteMap;
+  categoricalPalettes?: CategoricalPaletteMap;
   numCategories: number;
   selectedPalette: Color[];
   onChangePalette: (newPalette: Color[]) => void;
@@ -26,6 +34,8 @@ type ColorRampSelectorProps = {
 const defaultProps: Partial<ColorRampSelectorProps> = {
   disabled: false,
   useCategoricalPalettes: false,
+  colorRamps: DEFAULT_COLOR_RAMPS,
+  categoricalPalettes: DEFAULT_CATEGORICAL_PALETTES,
 };
 
 /**
